[export] [channelz] Property list: Separate out value conversion from property setting
This will allow open extensibility of the set of types available for property list (we no longer need to add them to the PropertyList type itself), and allow for other shaped types to do the same set of conversions.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/771460018](http://cl/771460018)

PiperOrigin-RevId: 771460018